refactor: add favorites menu to login and get profile response

+++ domain/menu.go
@@ -13,7 +13,7 @@ type Menu struct {
 	Price       int            `gorm:"column:price" json:"price"`
 	Photo       string         `gorm:"column:photo" json:"photo"`
 	CategoryId  int            `gorm:"column:category_id" json:"category_id"`
-	Rating      float32        `gorm:"column:rating" json:"rating"`
+	Rating      int            `gorm:"column:rating" json:"rating"`
 	TotalReview int            `gorm:"column:total_review" json:"total_review"`
 	IsAvailable bool           `gorm:"column:is_available" json:"is_available"`
 	Description string         `gorm:"column:description" json:"description"`
